% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matches-data.R
\docType{data}
\name{Matches}
\alias{Matches}
\title{Dataset Matches}
\format{
A data frame.
}
\source{
See references.
}
\usage{
data(Matches)
}
\description{
Includes match results and data for men's football Leagues in England, France,
Germany, Italy, Spain, The Netherlands, and Turkey.
}
\examples{
data(Matches)
mean(Matches$Goals90Home)
table(Matches$League)
}
\references{
elo: \url{http://clubelo.com/}
Market Values: \url{https://www.transfermarkt.com}
Bookmaker Odds: \url{https://www.oddsportal.com}
match data: \url{https://www.kicker.de}
}
\keyword{Data}
\keyword{Football}
\keyword{Soccer}
